+++ CHANGELOG.md
@@ -168,6 +168,7 @@ $form-modal-transition:         visibility $form-modal-transition-duration $form
 
 ### FIXES
 
+- Trigger Class on Scroll: IE11 fix.
 - Swiper Fullscreen Gallery: Closing fixes.
 - Video: re-init when softpage is loaded.
 - Colopicker JS Bug fixed.